third_party/WebKit/LayoutTests/svg/custom/SVGException-expected.txt

Tests the properties of the exception thrown by rotateFromVector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS e.toString() is "InvalidAccessError: A parameter or an operation was not supported by the underlying object."
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) should be [object DOMExceptionPrototype]. Was [object Object].
PASS e.constructor.toString() is "function DOMException() { [native code] }"
PASS e.constructor is window.DOMException
PASS successfullyParsed is true

TEST COMPLETE
